requirements.txt
psutil
rich
requests
pyserial
pynmea2